Create GenerateUniqueBinaryString.c
Google Phone Interview Question:
Given a single string containing any of 1's, 0's and ?'s (e.g. "10?10?1", "00???11", "???", etc) as an input, and print a list of strings containing all the unique binary permutations of the input string.

Review: Suppose there are N ?'s in the input string and the string length is L, so there are 2^N unique permutations. Also, we need to loop the entire string to find the position of ?, therefore, the time complexity is O(L + 2^N). Usually, 2^N >> L, so it's O(2^N), which leads a solution of divide and conquer.